gov_crawler/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
pyyaml>=6.0
python-dateutil>=2.8.0
urllib3>=2.0.0
lxml>=4.9.0
aiohttp>=3.9.0
tenacity>=8.2.0
